feat: notification settings split for weekly
